Update App.tsx
Reverting back to working version with autoCAD model and markers. In preparation of next step to switch netlify environmental variables to concordbridge5poc iModel with reality model and no CAD or markers.

I just recorded a youtube video of the autoCAD and marker version of the app. Burning the boats! Reality Model with map, terrain, and markers or bust!